build-tester.sh: seperate html generation

+++ script/build-tester.sh
@@ -23,9 +23,14 @@ BUILD_FOLDER="$SOURCE_FOLDER/.pio/build"
 CONFIG_FILE="$SOURCE_FOLDER/src/config/config.h"
 TARGETS_FILE="$SCRIPT_FOLDER/targets.json"
 
-BRANCH=$DRONE_BRANCH
+BRANCH=${GITHUB_REF#refs/heads/}
 if [ -z "$BRANCH" ]; then
   BRANCH="$(git rev-parse --abbrev-ref HEAD)"
+fi
+
+COMMIT=$GITHUB_SHA
+if [ -z "$COMMIT" ]; then
+  COMMIT="$(git rev-parse HEAD  )"
 fi
 
 BUILD_PREFIX="quicksilver"
@@ -71,7 +76,6 @@ function build_target() {
     if pio run -e $TARGET_NAME; then 
       cp "$BUILD_FOLDER/$TARGET_NAME/firmware.hex" "$OUTPUT_FOLDER/$BUILD_NAME.hex"
       echo -e "\e[32mSuccessfully\e[39m built target $BUILD_NAME"
-      echo "<a class=\"list-group-item list-group-item-action\" href=\"$BUILD_NAME.hex\" download target=\"_blank\">$BUILD_NAME</a>" >> $OUTPUT_FOLDER/index.html
     else
       echo -e "\e[31mError\e[39m building target $BUILD_NAME"
       exit 1
@@ -79,7 +83,7 @@ function build_target() {
   done
 }
 
-function build_all_targets() {
+function generate_html() {
   cat <<-EOF > $OUTPUT_FOLDER/index.html
 <!doctype html>
 <html lang="en">
@@ -98,13 +102,26 @@ function build_all_targets() {
       <h1 class="display-5 mb-4 fw-bold">Quicksilver $BRANCH</h1>
       <div class="col-lg-6 mx-auto">
         <p class="lead mb-4">
-          Commit <a href="https://github.com/BossHobby/QUICKSILVER/commit/$DRONE_COMMIT">$DRONE_COMMIT</a>
+          Commit <a href="https://github.com/BossHobby/QUICKSILVER/commit/$COMMIT">$COMMIT</a>
         </p>
         <div class="list-group">
 EOF
 
   for target in $(jq -r '.[] | @base64' $TARGETS_FILE); do
-    build_target $target
+    target_get() {
+      echo $target | base64 --decode | jq -r "${1}"
+    }
+
+    TARGET_NAME="$(target_get '.name')"
+    for config in $(target_get '.configurations[] | @base64'); do
+      config_get() {
+        echo ${config} | base64 --decode | jq -r "${1}"
+      }
+      CONFIG_NAME="$(config_get '.name')"
+      BUILD_NAME="$BUILD_PREFIX.$TARGET_NAME.$CONFIG_NAME"
+    done
+
+    echo "<a class=\"list-group-item list-group-item-action\" href=\"$BUILD_NAME.hex\" download target=\"_blank\">$BUILD_NAME</a>" >> $OUTPUT_FOLDER/index.html
   done
 
   cat <<-EOF >> $OUTPUT_FOLDER/index.html
@@ -118,14 +135,33 @@ EOF
 EOF
 }
 
+function build_all_targets() {
+  for target in $(jq -r '.[] | @base64' $TARGETS_FILE); do
+    build_target $target
+  done
+}
+
+if [[ $# -eq 0 ]]; then
+  echo "missing argument"
+  exit 1
+fi 
+
 jq 'empty' ./script/targets.json
 
-rm -rf $OUTPUT_FOLDER
-mkdir $OUTPUT_FOLDER
+mkdir $OUTPUT_FOLDER || true
 
-if [[ $# -ne 1 ]]; then
-  build_all_targets
-else 
-  TARGET=$(jq -r ".[] | select(.name==\""$1"\") | @base64" $TARGETS_FILE)
-  build_target $TARGET
+if [ "$1" == "build" ]; then
+  if [[ $# -ne 2 ]]; then
+    rm -rf $OUTPUT_FOLDER
+    mkdir $OUTPUT_FOLDER
+
+    build_all_targets
+  else 
+    TARGET=$(jq -r ".[] | select(.name==\""$2"\") | @base64" $TARGETS_FILE)
+    build_target $TARGET
+  fi
+elif [ "$1" == "html" ]; then
+  generate_html
+else
+  echo "unknown argument $1"
 fi
